+++ StockFlowPro.Web/App_Data/mock-users.json
@@ -9,7 +9,7 @@
     "dateOfBirth": "1980-01-01T00:00:00Z",
     "age": 45,
     "isActive": true,
-    "createdAt": "2025-06-24T13:25:25.335794Z",
+    "createdAt": "2025-06-24T14:02:57.7017717Z",
     "updatedAt": null,
     "role": "Admin",
     "passwordHash": "lhYqmKfEnhdQZTEU1c/w9JRG1B\u002BA9RmMvwNrMqDo8wM=:550e8400-e29b-41d4-a716-446655440000"
@@ -24,7 +24,7 @@
     "dateOfBirth": "1985-05-15T00:00:00Z",
     "age": 40,
     "isActive": true,
-    "createdAt": "2025-06-24T13:25:25.3466228Z",
+    "createdAt": "2025-06-24T14:02:57.7192715Z",
     "updatedAt": null,
     "role": "Admin",
     "passwordHash": "9uY5g\u002B2Xn6RzY56rnJDwCcEECPdHqCwyYbugJJ\u002BeJxc=:550e8400-e29b-41d4-a716-446655440001"
@@ -39,7 +39,7 @@
     "dateOfBirth": "1990-08-22T00:00:00Z",
     "age": 35,
     "isActive": true,
-    "createdAt": "2025-06-24T13:25:25.3467553Z",
+    "createdAt": "2025-06-24T14:02:57.7194824Z",
     "updatedAt": null,
     "role": "Manager",
     "passwordHash": "DlDCngVJc6Ywv1XU0AmJitu32FCM75j2EaTdX/k2LTI=:550e8400-e29b-41d4-a716-446655440002"
@@ -54,7 +54,7 @@
     "dateOfBirth": "1992-12-10T00:00:00Z",
     "age": 33,
     "isActive": true,
-    "createdAt": "2025-06-24T13:25:25.3467653Z",
+    "createdAt": "2025-06-24T14:02:57.7194997Z",
     "updatedAt": null,
     "role": "User",
     "passwordHash": "hcOYa4Y0UUm8qHILwKbuHqeH9CJZCYsGoEEvF6xMxnI=:550e8400-e29b-41d4-a716-446655440003"
@@ -69,7 +69,7 @@
     "dateOfBirth": "1988-03-07T00:00:00Z",
     "age": 37,
     "isActive": false,
-    "createdAt": "2025-06-24T13:25:25.346782Z",
+    "createdAt": "2025-06-24T14:02:57.719508Z",
     "updatedAt": null,
     "role": "User",
     "passwordHash": "8\u002BupOWwjzjPmhLwLcvIQqlO3FlqJed0q0pOkYyYU7sA=:550e8400-e29b-41d4-a716-446655440004"
