+++ StockFlowPro.Web/Controllers/Api/ProductsController.cs
@@ -5,6 +5,9 @@ using Microsoft.AspNetCore.Mvc;
 using StockFlowPro.Application.Commands.Products;
 using StockFlowPro.Application.DTOs;
 using StockFlowPro.Application.Queries.Products;
+using StockFlowPro.Shared.Extensions;
+using StockFlowPro.Shared.Models;
+using StockFlowPro.Shared.Constants;
 
 namespace StockFlowPro.Web.Controllers.Api;
 
@@ -23,25 +26,186 @@ public class ProductsController : ControllerBase
     }
 
     [HttpGet]
-    public async Task<ActionResult<IEnumerable<ProductDto>>> GetAllProducts(
+    public async Task<ActionResult<ApiResponse<IEnumerable<EnhancedProductDto>>>> GetAllProducts(
         [FromQuery] bool activeOnly = false,
         [FromQuery] bool inStockOnly = false,
         [FromQuery] bool lowStockOnly = false,
         [FromQuery] int lowStockThreshold = 10)
     {
-        Console.WriteLine($"[PRODUCT MANAGEMENT] Getting all products from database - ActiveOnly: {activeOnly}, InStockOnly: {inStockOnly}, LowStockOnly: {lowStockOnly}, LowStockThreshold: {lowStockThreshold}");
-        
-        var query = new GetAllProductsQuery 
-        { 
-            ActiveOnly = activeOnly,
-            InStockOnly = inStockOnly,
-            LowStockOnly = lowStockOnly,
-            LowStockThreshold = lowStockThreshold
+        try
+        {
+            Console.WriteLine($"[PRODUCT MANAGEMENT] Getting all products from database - ActiveOnly: {activeOnly}, InStockOnly: {inStockOnly}, LowStockOnly: {lowStockOnly}, LowStockThreshold: {lowStockThreshold}");
+            
+            var query = new GetAllProductsQuery 
+            { 
+                ActiveOnly = activeOnly,
+                InStockOnly = inStockOnly,
+                LowStockOnly = lowStockOnly,
+                LowStockThreshold = lowStockThreshold
+            };
+            var products = await _mediator.Send(query);
+            
+            // 🎯 ENHANCEMENT: Transform products with shared utilities
+            var enhancedProducts = products.Select(p => new EnhancedProductDto
+            {
+                // Original properties
+                Id = p.Id,
+                Name = p.Name,
+                CostPerItem = p.CostPerItem,
+                NumberInStock = p.NumberInStock,
+                TotalValue = p.TotalValue,
+                IsActive = p.IsActive,
+                IsInStock = p.IsInStock,
+                IsLowStock = p.IsLowStock,
+                CreatedAt = p.CreatedAt,
+                
+                // 🚀 ENHANCED: Formatted properties using shared utilities
+                FormattedName = p.Name.ToTitleCase(),
+                FormattedPrice = p.CostPerItem.ToCurrency(),
+                FormattedTotalValue = p.TotalValue.ToCurrency(),
+                FormattedTotalValueShort = p.TotalValue.ToShortFormat(),
+                StockDisplay = $"{p.NumberInStock} units",
+                CreatedDisplay = p.CreatedAt.ToDisplayFormat(),
+                CreatedFriendly = p.CreatedAt.ToFriendlyString(),
+                StockStatus = GetStockStatusText(p.IsInStock, p.IsLowStock),
+                StockStatusBadge = GetStockStatusBadge(p.IsInStock, p.IsLowStock),
+                ActiveStatusBadge = p.IsActive ? "Active" : "Inactive",
+                PriceRange = GetPriceRange(p.CostPerItem),
+                StockLevel = GetStockLevel(p.NumberInStock)
+            }).ToList();
+            
+            Console.WriteLine($"[PRODUCT MANAGEMENT] Retrieved and enhanced {enhancedProducts.Count} products from database");
+            
+            return Ok(ApiResponse<IEnumerable<EnhancedProductDto>>.SuccessResult(
+                enhancedProducts, 
+                $"Successfully retrieved {enhancedProducts.Count} products with enhanced formatting"));
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine($"[PRODUCT MANAGEMENT] Error retrieving products: {ex.Message}");
+            return BadRequest(ApiResponse<IEnumerable<EnhancedProductDto>>.ErrorResult(
+                "Failed to retrieve products", ex.Message));
+        }
+    }
+
+    [HttpGet("dashboard-stats")]
+    [Authorize(Roles = "Manager,Admin")]
+    public async Task<ActionResult<ApiResponse<EnhancedDashboardStats>>> GetDashboardStats()
+    {
+        try
+        {
+            Console.WriteLine("[PRODUCT MANAGEMENT] Getting enhanced dashboard statistics from database");
+            
+            var allProductsQuery = new GetAllProductsQuery { ActiveOnly = true };
+            var lowStockQuery = new GetAllProductsQuery { ActiveOnly = true, LowStockOnly = true };
+
+            var allProducts = await _mediator.Send(allProductsQuery);
+            var lowStockProducts = await _mediator.Send(lowStockQuery);
+            var outOfStockProducts = allProducts.Where(p => !p.IsInStock);
+
+            var totalValue = allProducts.Sum(p => p.TotalValue);
+            var totalProducts = allProducts.Count();
+            var lowStockCount = lowStockProducts.Count();
+            var outOfStockCount = outOfStockProducts.Count();
+
+            // 🚀 ENHANCED: Dashboard stats with shared utilities formatting
+            var enhancedStats = new EnhancedDashboardStats
+            {
+                // Raw values
+                TotalProducts = totalProducts,
+                TotalValue = totalValue,
+                LowStockCount = lowStockCount,
+                OutOfStockCount = outOfStockCount,
+                InStockCount = allProducts.Count(p => p.IsInStock),
+                
+                // 🎯 ENHANCED: Formatted values using shared utilities
+                TotalProductsFormatted = totalProducts.ToString("N0"),
+                TotalValueFormatted = totalValue.ToCurrency(),
+                TotalValueShort = totalValue.ToShortFormat(),
+                LowStockPercentage = ((decimal)lowStockCount).PercentageOf((decimal)totalProducts).ToString("F1") + "%",
+                OutOfStockPercentage = ((decimal)outOfStockCount).PercentageOf((decimal)totalProducts).ToString("F1") + "%",
+                InStockPercentage = ((decimal)allProducts.Count(p => p.IsInStock)).PercentageOf((decimal)totalProducts).ToString("F1") + "%",
+                LastUpdated = DateTime.Now.ToFriendlyString(),
+                LastUpdatedFull = DateTime.Now.ToDisplayFormat(),
+                
+                // Additional insights
+                AverageProductValue = totalProducts > 0 ? (totalValue / totalProducts).ToCurrency() : "$0.00",
+                HealthScore = CalculateInventoryHealthScore(totalProducts, lowStockCount, outOfStockCount),
+                StatusSummary = GetInventoryStatusSummary(totalProducts, lowStockCount, outOfStockCount)
+            };
+
+            Console.WriteLine($"[PRODUCT MANAGEMENT] Enhanced dashboard stats retrieved - Total: {enhancedStats.TotalProductsFormatted}, Value: {enhancedStats.TotalValueFormatted}, Health: {enhancedStats.HealthScore}%");
+            
+            return Ok(ApiResponse<EnhancedDashboardStats>.SuccessResult(
+                enhancedStats, 
+                "Enhanced dashboard statistics retrieved successfully"));
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine($"[PRODUCT MANAGEMENT] Error getting dashboard stats: {ex.Message}");
+            return BadRequest(ApiResponse<EnhancedDashboardStats>.ErrorResult(
+                "Failed to get enhanced dashboard stats", ex.Message));
+        }
+    }
+
+    // 🎯 HELPER METHODS using shared utilities
+    private static string GetStockStatusText(bool isInStock, bool isLowStock)
+    {
+        if (!isInStock) {return "Out of Stock";}
+        if (isLowStock) {return "Low Stock";}
+        return "In Stock";
+    }
+
+    private static string GetStockStatusBadge(bool isInStock, bool isLowStock)
+    {
+        if (!isInStock) {return "danger";}
+        if (isLowStock) {return "warning";}
+        return "success";
+    }
+
+    private static string GetPriceRange(decimal price)
+    {
+        return price switch
+        {
+            < 10 => "Budget",
+            < 50 => "Standard",
+            < 200 => "Premium",
+            _ => "Luxury"
         };
-        var products = await _mediator.Send(query);
+    }
+
+    private static string GetStockLevel(int stock)
+    {
+        return stock switch
+        {
+            0 => "Empty",
+            <= 5 => "Critical",
+            <= 20 => "Low",
+            <= 100 => "Good",
+            _ => "Excellent"
+        };
+    }
+
+    private static int CalculateInventoryHealthScore(int total, int lowStock, int outOfStock)
+    {
+        if (total == 0) {return 0;}
         
-        Console.WriteLine($"[PRODUCT MANAGEMENT] Retrieved {products.Count()} products from database");
-        return Ok(products);
+        var healthyStock = total - lowStock - outOfStock;
+        return (int)((decimal)healthyStock / total * 100);
+    }
+
+    private static string GetInventoryStatusSummary(int total, int lowStock, int outOfStock)
+    {
+        var healthScore = CalculateInventoryHealthScore(total, lowStock, outOfStock);
+        
+        return healthScore switch
+        {
+            >= 90 => "Excellent - Inventory levels are optimal",
+            >= 75 => "Good - Minor stock attention needed",
+            >= 50 => "Fair - Several items need restocking",
+            >= 25 => "Poor - Immediate restocking required",
+            _ => "Critical - Urgent inventory management needed"
+        };
     }
 
     [HttpGet("{id:guid}")]
@@ -157,31 +321,56 @@ public class ProductsController : ControllerBase
 
         return NoContent();
     }
+}
 
-    [HttpGet("dashboard-stats")]
-    [Authorize(Roles = "Manager,Admin")]
-    public async Task<ActionResult<object>> GetDashboardStats()
-    {
-        Console.WriteLine("[PRODUCT MANAGEMENT] Getting dashboard statistics from database");
-        
-        var allProductsQuery = new GetAllProductsQuery { ActiveOnly = true };
-        var lowStockQuery = new GetAllProductsQuery { ActiveOnly = true, LowStockOnly = true };
+// 🚀 ENHANCED DTOs with formatted properties
+public class EnhancedProductDto
+{
+    // Original properties
+    public Guid Id { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public decimal CostPerItem { get; set; }
+    public int NumberInStock { get; set; }
+    public decimal TotalValue { get; set; }
+    public bool IsActive { get; set; }
+    public bool IsInStock { get; set; }
+    public bool IsLowStock { get; set; }
+    public DateTime CreatedAt { get; set; }
+    
+    // 🎯 ENHANCED: Formatted properties using shared utilities
+    public string FormattedName { get; set; } = string.Empty;
+    public string FormattedPrice { get; set; } = string.Empty;
+    public string FormattedTotalValue { get; set; } = string.Empty;
+    public string FormattedTotalValueShort { get; set; } = string.Empty;
+    public string StockDisplay { get; set; } = string.Empty;
+    public string CreatedDisplay { get; set; } = string.Empty;
+    public string CreatedFriendly { get; set; } = string.Empty;
+    public string StockStatus { get; set; } = string.Empty;
+    public string StockStatusBadge { get; set; } = string.Empty;
+    public string ActiveStatusBadge { get; set; } = string.Empty;
+    public string PriceRange { get; set; } = string.Empty;
+    public string StockLevel { get; set; } = string.Empty;
+}
 
-        var allProducts = await _mediator.Send(allProductsQuery);
-        var lowStockProducts = await _mediator.Send(lowStockQuery);
-        var outOfStockProducts = allProducts.Where(p => !p.IsInStock);
-
-        var stats = new
-        {
-            TotalProducts = allProducts.Count(),
-            TotalValue = allProducts.Sum(p => p.TotalValue),
-            LowStockCount = lowStockProducts.Count(),
-            OutOfStockCount = outOfStockProducts.Count(),
-            InStockCount = allProducts.Count(p => p.IsInStock)
-        };
-
-        Console.WriteLine($"[PRODUCT MANAGEMENT] Dashboard stats retrieved - Total: {stats.TotalProducts}, In Stock: {stats.InStockCount}, Low Stock: {stats.LowStockCount}, Out of Stock: {stats.OutOfStockCount}");
-        
-        return Ok(stats);
-    }
+public class EnhancedDashboardStats
+{
+    // Raw values
+    public int TotalProducts { get; set; }
+    public decimal TotalValue { get; set; }
+    public int LowStockCount { get; set; }
+    public int OutOfStockCount { get; set; }
+    public int InStockCount { get; set; }
+    
+    // 🎯 ENHANCED: Formatted values
+    public string TotalProductsFormatted { get; set; } = string.Empty;
+    public string TotalValueFormatted { get; set; } = string.Empty;
+    public string TotalValueShort { get; set; } = string.Empty;
+    public string LowStockPercentage { get; set; } = string.Empty;
+    public string OutOfStockPercentage { get; set; } = string.Empty;
+    public string InStockPercentage { get; set; } = string.Empty;
+    public string LastUpdated { get; set; } = string.Empty;
+    public string LastUpdatedFull { get; set; } = string.Empty;
+    public string AverageProductValue { get; set; } = string.Empty;
+    public int HealthScore { get; set; }
+    public string StatusSummary { get; set; } = string.Empty;
 }
